check npm version before installation
