+++ lib/setup.js
@@ -379,9 +379,9 @@ function setupLinux(callback) {
 var gIsSudo = false;
 
 function npmInstall(packet, isSudo, callback) {
-	var exec = require('child_process').exec;
+    var exec = require('child_process').exec;
     var platform = require('os').platform();
-	// install from npm
+    // install from npm
     // Calculate directory
     var parts = __dirname.replace(/\\/g, '/').split('/');
     parts.splice(parts.length - 3, 3);
@@ -391,15 +391,15 @@ function npmInstall(packet, isSudo, callback) {
 
     var cmd = '';
     if (platform === 'linux' || platform === 'darwin') {
-        cmd = 'npm install ' + packet + ' --save --prefix ' + _path;
+        cmd = 'npm install ' + packet + ' --production --save --prefix ' + _path;
         if (isSudo) cmd = 'sudo ' + cmd;
     } else {
-        cmd = 'npm install ' + packet + ' --save --prefix "' + _path + '"';
+        cmd = 'npm install ' + packet + ' --production --save --prefix "' + _path + '"';
     }
     console.log(cmd);
     var child = exec(cmd);
-	child.stderr.pipe(process.stdout);
-	child.on('exit', function () {
+    child.stderr.pipe(process.stdout);
+    child.on('exit', function () {
         // Check if exist
         if ((platform === 'linux' || platform === 'darwin') && !isSudo && !fs.existsSync(_path + '/node_modules/' + packet)) {
             console.log('Cannot install as normal user. Try sudo...');
@@ -448,8 +448,8 @@ function setup(callback) {
 
     // install io-broker.js-controller
     npmInstall('iobroker.discovery', false, function () {
-		npmInstall('iobroker.admin', false, function () {
-			npmInstall('https://github.com/AlCalzone/ioBroker.js-controller/tarball/npm5', false, function () {
+        npmInstall('iobroker.admin', false, function () {
+            npmInstall('https://github.com/AlCalzone/ioBroker.js-controller/tarball/npm5', false, function () {
 
             if (!fs.existsSync(rootDir + 'iobroker-data/iobroker.json')) {
                     if (fs.existsSync(jsDir + 'conf/iobroker-dist.json')) {
@@ -484,8 +484,8 @@ function setup(callback) {
                     console.log('Non-critical error: ' + e.message);
                 }
             });
-		});
-	});
+        });
+    });
 }
 
 function setChmod(callback) {
@@ -512,71 +512,4 @@ function setChmod(callback) {
     }
 }
 
-function checkNpmVersion(callback) {
-    // Get npm version
-    try {
-        var nodeVersion = semver.valid(process.version);
-
-        var child = child_process.exec('npm -v');
-        var npmVersion;
-        var outResp = '';
-        var errResp = '';
-        child.stdout.on('data', function (buffer) {
-            outResp += buffer.toString();
-        });
-        child.stderr.on('data', function (buffer) {
-            errResp += buffer.toString();
-        });
-        child.stdout.on('end', function () {
-            if (outResp) {
-                console.log('NPM version: ' + outResp);
-                npmVersion = semver.valid(outResp.split('\n')[0], true);
-            }
-            if (errResp) {
-                console.error('Error trying to check npm version: ' + errResp);
-            }
-
-            if (!npmVersion) {
-                console.error('!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
-                console.error('Aborting install because the npm version could not be checked!');
-                console.error('Please check that npm is installed correctly.');
-                console.error('Use "npm install -g npm@4" or "npm install -g npm@>=5.7.1" to install a supported version.');
-                console.error('You need to make sure to repeat this step after installing an update to NodeJS and/or npm');
-                console.error('!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
-                processExit(1);
-                return;
-            }
-
-            if (semver.gte(npmVersion, "5.0.0") && semver.lt(npmVersion, "5.7.1")) {
-                console.error('!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
-                console.error('NPM 5 is only supported starting with version 5.7.1!');
-                console.error('Please use "npm install -g npm@4" to downgrade npm to 4.x or ');
-                console.error('use "npm install -g npm@>=5.7.1" to install a supported version of npm 5!')
-                console.error('You need to make sure to downgrade again with the above command after you');
-                console.error('You need to make sure to repeat this step after installing an update to NodeJS and/or npm');
-                console.error('!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
-                processExit(1);
-                return;
-            }
-            if (nodeVersion && semver.lt(nodeVersion, "4.0.0")) {
-                console.error('!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
-                console.error('ioBroker needs at least nodejs 4.x. You have installed ' + nodeVersion);
-                console.error('Please update your nodejs version!');
-                console.error('!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!');
-                processExit(1);
-                return;
-            }
-            callback();
-        });
-    }
-    catch (e) {
-        console.error('Could not check npm version. Assuming that correct version is installed.');
-    }
-}
-
-
-checkNpmVersion(function() {
-    setup(function (code) {
-    	processExit(code);
-    });
-});
+setup(processExit);
